+++ src/main/java/zone/moddev/mc/ironagefurniture/Ironagefurniture.java
@@ -6,7 +6,9 @@ import java.util.Map;
 import zone.moddev.mc.ironagefurniture.api.entity.Seat;
 import zone.moddev.mc.ironagefurniture.api.CreativeModeBreakTracker;
 import zone.moddev.mc.ironagefurniture.api.entity.EntityThrownLavaLamp;
+import zone.moddev.mc.ironagefurniture.api.tile.TileEntityPaddedBench;
 import zone.moddev.mc.ironagefurniture.init.BlockInitialiser;
+import zone.moddev.mc.ironagefurniture.init.ClientModelInitialiser;
 import zone.moddev.mc.ironagefurniture.init.ItemInitialiser;
 import zone.moddev.mc.ironagefurniture.lib.util.MMDCreativeTab;
 
@@ -19,6 +21,7 @@ import net.minecraftforge.fml.common.Mod.EventHandler;
 import net.minecraftforge.fml.common.event.FMLInitializationEvent;
 import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
 import net.minecraftforge.fml.common.registry.EntityRegistry;
+import net.minecraftforge.fml.common.registry.GameRegistry;
 import net.minecraftforge.common.MinecraftForge;
 
 @Mod(modid = Ironagefurniture.MODID, version = Ironagefurniture.VERSION)
@@ -53,7 +56,15 @@ public class Ironagefurniture
     public void preInit(FMLPreInitializationEvent event) {
     	IronAgeFurnitureConfiguration.init(event);
         MinecraftForge.EVENT_BUS.register(new CreativeModeBreakTracker());
+		GameRegistry.registerTileEntity(TileEntityPaddedBench.class,
+				new ResourceLocation(MODID, "padded_bench_colour"));
+		if (event.getSide().isClient()) {
+			ClientModelInitialiser.registerPaddedBenchModels();
+		}
     	BlockInitialiser.init();
+		if (event.getSide().isClient()) {
+			ClientModelInitialiser.registerPaddedBenchItemModels();
+		}
     	
     }
 }