+++ src/main/java/zone/moddev/mc/ironagefurniture/init/ItemInitialiser.java
@@ -1,6 +1,7 @@
 package zone.moddev.mc.ironagefurniture.init;
 
 import zone.moddev.mc.ironagefurniture.Ironagefurniture;
+import zone.moddev.mc.ironagefurniture.api.Items.ItemBlockPaddedBench;
 
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.renderer.block.model.ModelResourceLocation;
@@ -12,6 +13,9 @@ public class ItemInitialiser {
 
 		for(String name : Ironagefurniture.ItemRegistry.keySet()){
 			Item i = Ironagefurniture.ItemRegistry.get(name);
+			if (i instanceof ItemBlockPaddedBench) {
+				continue;
+			}
     		Minecraft.getMinecraft().getRenderItem().getItemModelMesher()
     				.register(i, 0, new ModelResourceLocation(Ironagefurniture.MODID + ":" + name, "inventory"));
     	}